+++ ptc_runner_launcher/release_config.exs
@@ -4,9 +4,11 @@
   macos_deployment_target: "15.0",
   precompiled_targets: %{
     {:unix, :darwin} => %{
-      "aarch64-apple-darwin" => {"gcc", "g++", "<%= cc %> -arch arm64", "<%= cxx %> -arch arm64"}
+      "aarch64-apple-darwin" => {"gcc", "g++", "<%= cc %> -arch arm64", "<%= cxx %> -arch arm64"},
+      "x86_64-apple-darwin" => {"gcc", "g++", "<%= cc %> -arch x86_64", "<%= cxx %> -arch x86_64"}
     },
     {:unix, :linux} => %{
+      "aarch64-linux-gnu" => "aarch64-linux-gnu-",
       "x86_64-linux-gnu" => "x86_64-linux-gnu-"
     }
   }